refactor(#792): split runtime_dsp_tests into base + extended battery (sweep)
runtime_dsp_tests.rs 869 -> 391, + runtime_dsp_tests_extended.rs 485 (both under
cap; off the debt list). Two-file test-split pattern: the shared signal helpers
(out_stage/sine/impulse) become pub(super) in the base file, and the extended
structural battery imports them via super::tests. Base helpers used only by the
base suite stay private. engine tests (36 result groups) pass, zero warnings.
